Rewrite the Loutre using Dioxus
At first, this was supposed to be a very simple application. However, as
features were added, the technical debt growth a lot and the code
because excessively difficult to maintain.

An architectural rewrite was in order, and with it came the question of
keeping egui or switching to something else. After evaluating all the
pros and cons, it was decided to drop egui in favor of Dioxus, which
allows for a better architectural design.

The rewrite is not finished yet and currently has no deadline.

https://dioxuslabs.com/
https://docs.rs/dioxus/latest/dioxus/
https://github.com/DioxusLabs/dioxus/